moves pattern command to site editor main navigation
